Move tests to integration and unit test folders
Sort out each test to the matching folder.
